solves issue #29 + refactoring
The camera can now be moved relative to the scene by dragging with the right mouse button, both in orbital as well as WASD mode. In my opinion using the right mouse button instead of the middle one results in a better user experience. When dragging with the middle mouse button it can easily happen that also the wheel moves a little bit.

Please try the feature out, do a code review and let me know about improvement ideas. I thought of making the dragging smoother by adding some momentum and ignoring negligible changes in the mouse position. But I suggest we leave that for another issue.